+++ types/index.ts
@@ -1,7 +1,6 @@
 export * from './Enums';
 export * from './ApiRequests';
 export * from './ApiResponses';
-export * from './ApiParams';
 
 export type File = Express.Multer.File;
 export type Uuid = string;
